backend/requirements.txt
pdf2image == 1.10.0
pytesseract == 0.3.0
numpy
pandas
Pillow
fastapi[all]
pytest